Minimize file-open suppression
This update allows hovered editors to be included in normal
recent-file tracking, provided that you either manually focused
the popup (not autofocused) or opened another popup from it and
then returned focus to it.  If a popup is autofocused it will
not be treated as the file being opened for purposes of the
quick switcher or other plugins monitoring file-open.