Add a closeDelay setting
For improved accessibility, allow setting a different
delay for closing popovers than opening them. This makes
it easier for people with co-ordination challenges to
resize an unpinned popover without accidentally closing it.
Once a popover has been open for the close delay, it will
persist for the close delay before closing after the
cursor moves, unless a new popup is triggered sooner.

Also, adjusted the defaults and timing algorithms to more
closely match those of Obsidian core, which waits 100ms
after a hover is triggered to begin loading the file, and
shows the popover after 300ms (200ms after the file open
begins), and to ensure that even with a long close delay,
popups are not indefinitely cascaded for the same parent.