Improved compatibility with Obsidian 0.15.6
This version should fix some issues with opening things in
Hover Editors in Obsidian 0.15.6.  Because of the invasive
changes required to make these things work, this is being
released as a beta first.  Please report any issues!

(Also, if you are using Pane Relief with this beta, be sure to
update it to 0.2.0 or higher first.)