Fix #227 - Fold state being lost on editor hide
When hovering a link, but *not* long enough to open it,
closing the not-yet-shown pane would cause Obsidian to try
to save the fold state, which was not yet loaded because
the pane wasn't shown.  This change blocks Obsidian from
saving fold state when Hover Editor is closing a pane that
was never shown.  This should prevent fold state being lost
in such cases.