Add settings for footnotes, subheads, and blocks
You can now configure Hover Editor to *not* replace the built-in
preview functions for links to footnotes, subheadings, and/or
blocks.

By default, blocks and footnotes will now be displayed
with the built-in preview (with Obsidian 1.6 and higher),
but subheadings will still be displayed in Hover Editor.
(For older Obsidians, the default will be to use Hover Editor
for everything, since before 1.6 the built-in preview didn't
support editing.)

You can change these settings from the defaults using the plugin's
settings page.